Get rid of charging reason to make life easier with JSON
Unfortunately, the situation with JSON deserialization errors is getting
worse. I knew it, so I should have followed another path. Let me say it
once more:

Never, ever use string-based protocols/formats as your base.

Circa 2009: https://twitter.com/loverdos/status/5761681752

The bad news is that I do not have time to fix it in the next couple of
days due to a pending demo. Charging reason is not essential for the
demo, so I ditch it and put a reminder here so as to probably
reintroduce it in the future.